refactor: rename `tile-layer.tsx` into `TileLayer.tsx`
